solve linters
